Added dynamic visibility and enablement of certain UI elements
Pulse Configuration tab will now only be visible for a pulsed reactor.
Component detail buttons (the "a" and "i" above each component in the grid) can be hidden with an option in the new "Advanced" tab.
Added options to disable component palette buttons for GT 5.08 and GT 5.09 components (and to offer to remove such components from the current design if there are any)